requirements.txt
flask_restx
flask_script
flask_mongoengine
flask_redis
flask_testing
